Allow everybody to be selected as Chancellor after Anarchy
If you check the rulebook, you'll see that after the Anarchy, everybody is elegible as Chancellor.
Actually, this is not allowed by the bot. As it is controlled by the game.board.state. predient and chancellor, turning them into "None" will reflect properly the actual state of the game (during Anrchy, there are no president or chancellor). Then, everybody will be elegible for chancellor (except for the president) just as the start of the game.